Reverted back to r605, the working revision before the merge.

+++ src/game/bg_misc.c
@@ -1195,7 +1195,7 @@ static const classAttributes_t bg_classList[ ] =
     ( 1 << S1 )|( 1 << S2 )|( 1 << S3 ),            //int  stages
     0,                                              //int     health;
     0.0f,                                           //float   fallDamage;
-    0.0f,                                           //float   regenRate;
+    0,                                              //int     regenRate;
     0,                                              //int     regenRateMsec;
     0,                                              //int     abilities;
     WP_NONE,                                        //weapon_t  startWeapon
@@ -1223,7 +1223,7 @@ static const classAttributes_t bg_classList[ ] =
     ( 1 << S1 )|( 1 << S2 )|( 1 << S3 ),            //int  stages
     ABUILDER_HEALTH,                                //int     health;
     0.2f,                                           //float   fallDamage;
-    ABUILDER_REGEN,                                 //float   regenRate;
+    ABUILDER_REGEN,                                 //int     regenRate;
     ABUILDER_REGEN_MSEC,                            //int     regenRateMsec;
     SCA_TAKESFALLDAMAGE|SCA_FOVWARPS|SCA_ALIENSENSE,//int     abilities;
     WP_ABUILD,                                      //weapon_t  startWeapon
@@ -1252,7 +1252,7 @@ static const classAttributes_t bg_classList[ ] =
     ( 1 << S2 )|( 1 << S3 ),                        //int  stages
     ABUILDER_UPG_HEALTH,                            //int     health;
     0.2f,                                           //float   fallDamage;
-    ABUILDER_UPG_REGEN,                             //float   regenRate;
+    ABUILDER_UPG_REGEN,                             //int     regenRate;
     ABUILDER_UPG_REGEN_MSEC,                        //int     regenRateMsec;
     SCA_TAKESFALLDAMAGE|SCA_FOVWARPS|SCA_WALLCLIMBER|SCA_ALIENSENSE,    //int     abilities;
     WP_ABUILD2,                                     //weapon_t  startWeapon
@@ -1280,7 +1280,7 @@ static const classAttributes_t bg_classList[ ] =
     ( 1 << S1 )|( 1 << S2 )|( 1 << S3 ),            //int  stages
     LEVEL0_HEALTH,                                  //int     health;
     0.0f,                                           //float   fallDamage;
-    LEVEL0_REGEN,                                   //float   regenRate;
+    LEVEL0_REGEN,                                   //int     regenRate;
     LEVEL0_REGEN_MSEC,                              //int     regenRateMsec;
     SCA_WALLCLIMBER|SCA_FOVWARPS|SCA_ALIENSENSE,    //int     abilities;
     WP_ALEVEL0,                                     //weapon_t  startWeapon
@@ -1308,7 +1308,7 @@ static const classAttributes_t bg_classList[ ] =
     ( 1 << S1 )|( 1 << S2 )|( 1 << S3 ),            //int  stages
     LEVEL0_UPG_HEALTH,                              //int     health;
     0.0f,                                           //float   fallDamage;
-    LEVEL0_UPG_REGEN,                               //float   regenRate;
+    LEVEL0_UPG_REGEN,                               //int     regenRate;
     LEVEL0_UPG_REGEN_MSEC,                          //int     regenRateMsec;
     SCA_WALLCLIMBER|SCA_FOVWARPS|SCA_ALIENSENSE,    //int     abilities;
     WP_ALEVEL0_UPG,                                 //weapon_t  startWeapon
@@ -1327,6 +1327,18 @@ static const classAttributes_t bg_classList[ ] =
     { PCL_ALIEN_LEVEL1, PCL_ALIEN_BUILDER0, PCL_NONE },       //int     children[ 3 ];
     LEVEL0_UPG_COST,                                //int     cost;
     LEVEL0_UPG_VALUE                                //int     value;
+//     "jumper",                                       //char    *modelname;
+//     0.2f,                                           //float   modelScale;
+//     "default",                                      //char    *skinname;
+//     0.3f,                                           //float   shadowScale;
+//     "alien_general_hud",                            //char    *hudname;
+//     { -15, -15, -15 },                              //vec3_t  mins;
+//     { 15, 15, 15 },                                 //vec3_t  maxs;
+//     { 15, 15, 15 },                                 //vec3_t  crouchmaxs;
+//     { -15, -15, -4 },                               //vec3_t  deadmins;
+//     { 15, 15, 4 },                                  //vec3_t  deadmaxs;
+//     -8.0f,                                          //float   zOffset
+//     0, 0,                                           //int     viewheight, crouchviewheight;
   },
   {
     PCL_ALIEN_LEVEL1,                               //int     classnum;
@@ -1338,7 +1350,7 @@ static const classAttributes_t bg_classList[ ] =
     ( 1 << S1 )|( 1 << S2 )|( 1 << S3 ),            //int  stages
     LEVEL1_HEALTH,                                  //int     health;
     0.0f,                                           //float   fallDamage;
-    LEVEL1_REGEN,                                   //float   regenRate;
+    LEVEL1_REGEN,                                   //int     regenRate;
     LEVEL1_REGEN_MSEC,                              //int     regenRateMsec;
     SCA_FOVWARPS|SCA_WALLCLIMBER|SCA_ALIENSENSE,    //int     abilities;
     WP_ALEVEL1,                                     //weapon_t  startWeapon
@@ -1368,7 +1380,7 @@ static const classAttributes_t bg_classList[ ] =
     ( 1 << S2 )|( 1 << S3 ),                        //int  stages
     LEVEL1_UPG_HEALTH,                              //int     health;
     0.0f,                                           //float   fallDamage;
-    LEVEL1_UPG_REGEN,                               //float   regenRate;
+    LEVEL1_UPG_REGEN,                               //int     regenRate;
     LEVEL1_UPG_REGEN_MSEC,                          //int     regenRateMsec;
     SCA_FOVWARPS|SCA_WALLCLIMBER|SCA_ALIENSENSE,    //int     abilities;
     WP_ALEVEL1_UPG,                                 //weapon_t  startWeapon
@@ -1396,7 +1408,7 @@ static const classAttributes_t bg_classList[ ] =
     ( 1 << S1 )|( 1 << S2 )|( 1 << S3 ),            //int  stages
     LEVEL2_HEALTH,                                  //int     health;
     0.0f,                                           //float   fallDamage;
-    LEVEL2_REGEN,                                   //float   regenRate;
+    LEVEL2_REGEN,                                   //int     regenRate;
     LEVEL2_REGEN_MSEC,                              //int     regenRateMsec;
     SCA_WALLJUMPER|SCA_FOVWARPS|SCA_ALIENSENSE,     //int     abilities;
     WP_ALEVEL2,                                     //weapon_t  startWeapon
@@ -1425,7 +1437,7 @@ static const classAttributes_t bg_classList[ ] =
     ( 1 << S2 )|( 1 << S3 ),                        //int  stages
     LEVEL2_UPG_HEALTH,                              //int     health;
     0.0f,                                           //float   fallDamage;
-    LEVEL2_UPG_REGEN,                               //float   regenRate;
+    LEVEL2_UPG_REGEN,                               //int     regenRate;
     LEVEL2_UPG_REGEN_MSEC,                          //int     regenRateMsec;
     SCA_WALLJUMPER|SCA_FOVWARPS|SCA_ALIENSENSE,     //int     abilities;
     WP_ALEVEL2_UPG,                                 //weapon_t  startWeapon
@@ -1454,7 +1466,7 @@ static const classAttributes_t bg_classList[ ] =
     ( 1 << S1 )|( 1 << S2 )|( 1 << S3 ),            //int  stages
     LEVEL3_HEALTH,                                  //int     health;
     0.0f,                                           //float   fallDamage;
-    LEVEL3_REGEN,                                   //float   regenRate;
+    LEVEL3_REGEN,                                   //int     regenRate;
     LEVEL3_REGEN_MSEC,                              //int     regenRateMsec;
     SCA_FOVWARPS|SCA_ALIENSENSE,                    //int     abilities;
     WP_ALEVEL3,                                     //weapon_t  startWeapon
@@ -1483,7 +1495,7 @@ static const classAttributes_t bg_classList[ ] =
     ( 1 << S2 )|( 1 << S3 ),                                    //int  stages
     LEVEL3_UPG_HEALTH,                              //int     health;
     0.0f,                                           //float   fallDamage;
-    LEVEL3_UPG_REGEN,                               //float   regenRate;
+    LEVEL3_UPG_REGEN,                               //int     regenRate;
     LEVEL3_UPG_REGEN_MSEC,                          //int     regenRateMsec;
     SCA_FOVWARPS|SCA_ALIENSENSE,                    //int     abilities;
     WP_ALEVEL3_UPG,                                 //weapon_t  startWeapon
@@ -1513,7 +1525,7 @@ static const classAttributes_t bg_classList[ ] =
     ( 1 << S3 ),                                    //int  stages
     LEVEL4_HEALTH,                                  //int     health;
     0.0f,                                           //float   fallDamage;
-    LEVEL4_REGEN,                                   //float   regenRate;
+    LEVEL4_REGEN,                                   //int     regenRate;
     LEVEL4_REGEN_MSEC,                              //int     regenRateMsec;
     SCA_FOVWARPS|SCA_ALIENSENSE,                    //int     abilities;
     WP_ALEVEL4,                                     //weapon_t  startWeapon
@@ -1541,7 +1553,7 @@ static const classAttributes_t bg_classList[ ] =
     ( 1 << S3 ),                                    //int  stages
     LEVEL4_UPG_HEALTH,                              //int     health;
     0.0f,                                           //float   fallDamage;
-    LEVEL4_UPG_REGEN,                               //float   regenRate;
+    LEVEL4_UPG_REGEN,                               //int     regenRate;
     LEVEL4_UPG_REGEN_MSEC,                          //int     regenRateMsec;
     SCA_FOVWARPS|SCA_ALIENSENSE,                    //int     abilities;
     WP_ALEVEL4_UPG,                                 //weapon_t  startWeapon
@@ -1560,6 +1572,19 @@ static const classAttributes_t bg_classList[ ] =
     { PCL_NONE, PCL_NONE, PCL_NONE },               //int     children[ 3 ];
     LEVEL4_UPG_COST,                                //int     cost;
     LEVEL4_UPG_VALUE                                //int     value;
+    
+    
+//     1.0f,                                           //float   modelScale;
+//     "default",                                      //char    *skinname;
+//     2.0f,                                           //float   shadowScale;
+//     "alien_general_hud",                            //char    *hudname;
+//     { -32, -32, -22 },                              //vec3_t  mins;
+//     { 32, 32, 70 },                                 //vec3_t  maxs;
+//     { 32, 32, 70 },                                 //vec3_t  crouchmaxs;
+//     { -32, -32, -34 },                               //vec3_t  deadmins;
+//     { 32, 32, 34 },                                  //vec3_t  deadmaxs;
+//     0.0f,                                           //float   zOffset
+//     64, 64,                                         //int     viewheight, crouchviewheight;
   },
   {
     PCL_HUMAN,                                      //int     classnum;
@@ -1568,7 +1593,7 @@ static const classAttributes_t bg_classList[ ] =
     ( 1 << S1 )|( 1 << S2 )|( 1 << S3 ),            //int  stages
     100,                                            //int     health;
     1.0f,                                           //float   fallDamage;
-    0.0f,                                           //float   regenRate;
+    0,                                              //int     regenRate;
     0,                                              //int     regenRateMsec;
     SCA_TAKESFALLDAMAGE|SCA_CANUSELADDERS,          //int     abilities;
     WP_NONE, //special-cased in g_client.c          //weapon_t  startWeapon
@@ -1595,7 +1620,7 @@ static const classAttributes_t bg_classList[ ] =
     ( 1 << S3 ),                                    //int  stages
     100,                                            //int     health;
     1.0f,                                           //float   fallDamage;
-    0.0f,                                           //float   regenRate;
+    0,                                              //int     regenRate;
     0,                                              //int     regenRateMsec;
     SCA_TAKESFALLDAMAGE|
       SCA_CANUSELADDERS,                            //int     abilities;
